changed the purpose of the bulmaField component
The purpose was to wrap as many control as we want and put them grouped by default.
I thought then than non-grouped component would behave fine with multiple controls.

By default, the bulmaField is not grouped and should contain only one label and one control.

Unless the `has-addons`attribute is added on the component, `is-grouped` is applied as a class on the field
=> can insert as many fields as necessary

if the `has-addons`attribute is added on the component, the `has-addons` class is added to the field
=> can insert as many fields as necessary