created new superRadioButton component
This component is conceived in a peculiar way:

- the .vue file contains a second component declaration, private to this superRadioButton : superOption
- the superRadioButton uses JSX templating to allow as many named slot as necessary
- the superOption also uses JSX templating for two reasons : compilation (see below for details) and because an event is automatically binded to all slot elements (and their children) that could be used for input (input or buttons) so that the superOption is activated if they are modified (without clicks)

Because the superOption is specific to this superRadioButton component and carries some specific CSS that can only display correctly in this context

Because this 'template' option needs to be compiled in JS to work. Webpack will be doing it for us in the exported component, but will ignore the "private" superOption component. The solution is to  use JSX (there is more details to add on this: https://codewithhugo.com/writing-multiple-vue-components-in-a-single-file/)