ruleEditor: added condensed menu for edit options
NOTE 1: didn't use bulma's dropdown component. Instead created a custom "dropnext" component

NOTE 2: added a 'forceCloseEditMenu' prop with a watcher that hide the menu if the prop is true.
also, each click on the dropdown's trigger button emits a 'openmenu' event to the columnSettingsEditor component. The key of the triggering rule will be stored in the 'ruleCurentlyOpenedEditorKey' data. This will trigger a prop change for all other rules

NOTE 3: each click on a option will close the menu + do the associated effect