highlight & ruleEditor: make room for a description
modified the layout of the rule sumup to allow for the addition of short description :

- moved down the h5 tag _inside_ the .rule-sumup container
- moved the "small-swatch" into a .icon-container
- added a grid layout on the .rule-sumup (1.75rem 1fr)
- every new item in .rule-sumup, except for the .icon-container, will be in the second column
- added a grid-row span 2 on the .icon-container (in prevision to multiline description)

NOTE: ruleEditor currently do not have an icon-container. Could add one for action type icon.
NOTE 2: ruleEditor and highlightEditor have some common code (that why I had to modify both) : should create a UI component for it.